Make no-slip implementation GPU friendly and satisfy unit tests.
