+++ src/edge_kernels/CMakeLists.txt
@@ -9,7 +9,6 @@ target_sources(nalu PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/WallDistEdgeSolverAlg.C
   ${CMAKE_CURRENT_SOURCE_DIR}/MomentumEdgePecletAlg.C
   ${CMAKE_CURRENT_SOURCE_DIR}/StreletsUpwindEdgeAlg.C
-  ${CMAKE_CURRENT_SOURCE_DIR}/MomentumABLWallFuncMaskUtil.C
 
   # Face/edge BC kernels
   ${CMAKE_CURRENT_SOURCE_DIR}/ContinuityOpenEdgeKernel.C
